refactor(utils.model_pydantic): Move ImageMetadata and GroupMetadata to utils.model_pydantic
